refactor: union doctors query logic to single module
