--- src/doctor/helpers/sorting/sort-by-name-strategy.ts
@@ -1,8 +1,0 @@
-import { ISortingStrategy } from '../../../common/types';
-import { DoctorModel } from '../../doctor-model';
-
-export class SortByNameStrategy implements ISortingStrategy {
-	sortDescening(doctors: DoctorModel[]): DoctorModel[] {
-		return doctors.sort((a, b) => a.userId.localeCompare(b.userId));
-	}
-}
